backend/requirements.txt
# LLM & Pipeline
google-generativeai>=0.7.0
langgraph>=0.2.0
langchain-core>=0.2.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Validation
pydantic>=2.7.0

# PDF
PyMuPDF>=1.24.0

# Utils
python-dotenv>=1.0.0

# Dev / Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
